move Templates into its own library
